Layouts Save

+++ BusinessApi/Models/BaseDto.cs
@@ -3,9 +3,9 @@
     public class BaseDto
     {
         public int Czad { get; set; }
-        public int CreatedBy { get; set; }
+        public double CreatedBy { get; set; }
         public DateTime CreatedDate { get; set; }
-        public int UpdatedBy { get; set; }
+        public double UpdatedBy { get; set; }
         public DateTime UpdatedDate { get; set; }
     }
 }
